requirements.txt
fastapi
uvicorn
torch
openai
python-dotenv
pydantic
openm-core[cli]
